feat: 메인페이지 화면 추가 기존 success->main
