+++ nalchi-fe/src/components/authService.js
@@ -1,6 +1,7 @@
 import axios from 'axios';
 
 const API_BASE_URL = "http://localhost:8080/api/auth";
+// const API_BASE_URL = "http://3.38.246.140:8080/api/auth";
 
 const storeTokens = (tokens) => {
     if (tokens.accessToken && tokens.refreshToken) {
